edited calendarapp file: added functionality to switch between calendars depending on which one the user signs up with & created calendar's own folder structure
