+++ client/src/components/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import NavTabs from "./NavTabs";
 import ChatApp from './ChatApp';
-import CalendarApp from './CalendarApp';
+import CalendarApp from './calendar/CalendarApp';
 import DiaryApp from "./diary/DiaryApp";
 import SignupApp from './SignupApp';
