+++ client/src/components/calendar/CalendarApp.css
@@ -17,7 +17,7 @@
 }
 
 .rbc-event-content {
-  background-color: #4f81a0;
+  background-color: transparent;
 }
 
 .App {
