+++ client/src/components/calendar/index.js
@@ -1,0 +1,1 @@
+export {CalendarApp} from "./calendar";
